+++ components/PageLinkContainer/PageArrow.tsx
@@ -1,45 +1,41 @@
 import { AnimatePresence } from "framer-motion";
-import PageLinkModel from "../models/pageLink";
-import PageLinkAnimation from "../animations/PageLinkAnimation";
+import PageLinkModel from "../../models/pageLink";
+import ArrowAnimation from '../../animations/ArrowAnimation';
 
-const PageLink: React.FC<PageLinkModel> = ({ onHelloPage }) => (
+
+const PageArrow: React.FC<PageLinkModel> = ({ onHelloPage }) => (
   <>
-    {/* 
+  {/* 
       onHelloPage & !onHelloPage used to add/remove from DOM
       which triggers AnimatePresence to fire animation
-    */}
+  */}
     <AnimatePresence initial={false}>
       {onHelloPage && (
-        <span className="pageLink blue">
-          <PageLinkAnimation>about</PageLinkAnimation>
+        <span className="arrow red">
+          <ArrowAnimation>{"-->"}</ArrowAnimation>
         </span>
       )}
     </AnimatePresence>
     <AnimatePresence initial={false}>
       {!onHelloPage && (
-        <span className="pageLink red">
-          <PageLinkAnimation>hello</PageLinkAnimation>
+        <span className="arrow blue">
+          <ArrowAnimation>{"-->"}</ArrowAnimation>
         </span>
       )}
     </AnimatePresence>
 
     <style jsx>{`
-      .pageLink {
+      .arrow {
         position: absolute;
         display: block;
-        right: 60px;
+        right: 30px;
+        width: 25px;
         font-size: 0.9em;
         font-style: italic;
         cursor: pointer;
-      }
-
-      .linkText {
-        position: absolute;
-        display: inline;
-        z-index: 99;
       }
     `}</style>
   </>
 );
 
-export default PageLink;
+export default PageArrow;
